Module type Ast_builder.S_with_implicit_loc

type arrow_argument := Shim.arrow_argument

type arrow_result := Shim.arrow_result

type block_access := Shim.block_access

type modalities := Shim.Modalities.t

type modes := Shim.Modes.t

type include_kind := Shim.Include_kind.t

type index_kind = Shim.index_kind

type jkind_annotation := Shim.jkind_annotation

type unboxed_access := Shim.unboxed_access

module Pcstr_tuple_arg := Shim.Pcstr_tuple_arg

Modes

val ptyp_arrow : arrow_argument -> arrow_result -> Ppxlib_ast.Parsetree.core_type

Construct a Ptyp_arrow

val tarrow : arrow_argument list -> arrow_result -> Ppxlib_ast.Parsetree.core_type

Construct a multi-argument arrow type with the provided arguments and result.

raises [Invalid_argument]
if the input list is empty.

val tarrow_maybe : arrow_argument list -> Ppxlib_ast.Parsetree.core_type -> Ppxlib_ast.Parsetree.core_type

As tarrow, but will return the result if the input list is empty rather than erroring; this means the result type cannot have a mode annotation.

val pexp_let : Ppxlib_ast.Asttypes.mutable_flag -> Ppxlib_ast.Asttypes.rec_flag -> Ppxlib_ast.Parsetree.value_binding list -> Ppxlib_ast.Parsetree.expression -> Ppxlib_ast.Parsetree.expression

Pexp_let with let mutable support

val pexp_constraint : Ppxlib_ast.Parsetree.expression -> Ppxlib_ast.Parsetree.core_type option -> modes -> Ppxlib_ast.Parsetree.expression

Construct a Pexp_constraint with modes

val ppat_constraint : Ppxlib_ast.Parsetree.pattern -> Ppxlib_ast.Parsetree.core_type option -> modes -> Ppxlib_ast.Parsetree.pattern

Construct a Ppat_constraint with modes

val value_binding : pat:Ppxlib_ast.Parsetree.pattern -> expr:Ppxlib_ast.Parsetree.expression -> modes:modes -> Ppxlib_ast.Parsetree.value_binding

Contruct a value_binding with modes

val pcstr_tuple : (modalities * Ppxlib_ast.Parsetree.core_type) list -> Ppxlib_ast.Parsetree.constructor_arguments

Construct a Pcstr_tuple, a representation for the contents of a tupled variant constructor, that attaches the provided modalities to each field.

val psig_include : modalities:modalities -> Ppxlib_ast.Parsetree.include_description -> Ppxlib_ast.Parsetree.signature_item

Construct a Psig_include with modalities

val signature : ?modalities:modalities -> Ppxlib_ast.Parsetree.signature_item list -> Ppxlib_ast.Parsetree.signature

Construct a signature

val pmty_signature : Ppxlib_ast.Parsetree.signature -> Ppxlib_ast.Parsetree.module_type

val get_tuple_field_modalities : Pcstr_tuple_arg.t -> modalities * Ppxlib_ast.Parsetree.core_type

Splits a possibly-modality-annotated field of a tupled variant constructor into a pair of its modality and the unannotated field. If the resulting mode is None, then the field is returned unchanged.

val get_label_declaration_modalities : Ppxlib_ast.Parsetree.label_declaration -> modalities * Ppxlib_ast.Parsetree.label_declaration

Splits a possibly-modality-annotated label declaration into a pair of its modality and the unannotated label declaration. If the resulting modality is None, then the label declaration is returned unchanged.

val label_declaration : name:string Astlib.Location.loc -> mutable_:Ppxlib_ast.Asttypes.mutable_flag -> modalities:modalities -> type_:Ppxlib_ast.Parsetree.core_type -> Ppxlib_ast.Parsetree.label_declaration

val get_value_description_modalities : Ppxlib_ast.Parsetree.value_description -> modalities * Ppxlib_ast.Parsetree.value_description

val value_description : name:string Astlib.Location.loc -> type_:Ppxlib_ast.Parsetree.core_type -> modalities:modalities -> prim:string list -> Ppxlib_ast.Parsetree.value_description

val pcstr_tuple_arg : modalities:modalities -> type_:Ppxlib_ast.Parsetree.core_type -> Pcstr_tuple_arg.t

val include_infos : ?attrs:Ppxlib_ast.Parsetree.attributes -> kind:include_kind -> 'a -> 'a Ppxlib_ast.Parsetree.include_infos

val module_declaration : ?attrs:Ppxlib_ast.Parsetree.attributes -> ?modalities:modalities -> string option Astlib.Location.loc -> Ppxlib_ast.Parsetree.module_type -> Ppxlib_ast.Parsetree.module_declaration

val pmty_functor : ?attrs:Ppxlib_ast.Parsetree.attributes -> ?modes:modes -> Ppxlib_ast.Parsetree.functor_parameter -> Ppxlib_ast.Parsetree.module_type -> Ppxlib_ast.Parsetree.module_type

val pmod_constraint : ?attrs:Ppxlib_ast.Parsetree.attributes -> Ppxlib_ast.Parsetree.module_expr -> Ppxlib_ast.Parsetree.module_type option -> modes -> Ppxlib_ast.Parsetree.module_expr

N-ary functions

Many comments below make reference to the Jane Street compiler's treatment of function arity. These comments refer to a parsetree change made to upstream OCaml in https://github.com/ocaml/ocaml/pull/12236, but that Jane Street has mirrored internally already.

The treatment of arity can be summarized as follows:

In a previous version of OCaml, a function's runtime arity was inferred at a late stage of the compiler, after typechecking, where it fuses together nested lambdas.

In the new version of OCaml (both upstream OCaml after #12236 and the internal Jane Street compiler), a function's runtime arity is purely a syntactic notion: it's the number of parameters in a fun x1 ... xn -> body construct, with some special allowances for function cases.

Why is arity important? In native code, application sites of a function to n syntactic arguments will trigger a fast path (where arguments are passed in registers) only if the function's runtime arity is n.

As a result, ppxes must take more care than before to generate functions of the correct arity. Now, a nested function like fun x -> fun y -> e has arity 1 (returning still another function of arity 1) instead of arity 2. All bindings below that construct functions are documented as to the arity of the returned function.

Some examples of arity:

2-ary function: fun x y -> e

1-ary function returning 1-ary function: fun x -> fun y -> e

3-ary function: fun x y -> function P1 -> e1 | P2 -> e2

2-ary function returning 1-ary function: fun x y -> (function P1 -> e1 | P2 -> e2)

2-ary function returning 1-ary function: fun x -> function P1 -> function P2 -> e

Notably, unparenthesized function has a special meaning when used as a direct body of fun: the function becomes part of the arity of the outer fun. The same does not apply for multiple nested functions, even if they each have a single case; the nested functions are treated as unary. (See the last example.)

type function_param = Shim.Pexp_function.function_param

type function_constraint = Shim.Pexp_function.Function_constraint.t

type function_body = Shim.Pexp_function.function_body

module Latest : sig ... end

Avoid shadowing Ppxlib's AST builder.

val eabstract : ?coalesce_fun_arity:bool -> ?return_constraint:Ppxlib_ast.Parsetree.core_type -> Ppxlib_ast.Parsetree.pattern list -> Ppxlib_ast.Parsetree.expression -> Ppxlib_ast.Parsetree.expression

Create a function with unlabeled parameters and an expression body. Like Ppxlib.Ast_builder.eapply, but for constructing functions.

coalesce_fun_arity is relevant for the Jane Street compiler. By default, coalesce_fun_arity is true.

Suppose there is a call eabstract pats body ~coalesce_fun_arity

If colaesce_fun_arity is true, the arity of the returned function is the same as the arity of: add_fun_params (List.map params ~f:(Fun.param Nolabel)) body

If coalesce_fun_arity is false, then the arity of the returned function is the length of pats.

In other words, coalesce_fun_arity = true allows you to build up the arity of an already-constructed function rather than necessarily creating a new function.

val unary_function : ?attrs:Ppxlib_ast.Parsetree.attributes -> Ppxlib_ast.Parsetree.case list -> Ppxlib_ast.Parsetree.expression

unary_function cases is function <cases>. When used with the Jane Street compiler, the function's runtime arity is 1, so the fast path for function application happens only when application sites of the resulting function receive 1 argument. To create a function with multiple argument that pattern-matches on the last one, use add_param or add_params to add more parameters. Alternatively, use pexp_function to provide all parameters at once.

The attributes of the resulting expression will be the attrs argument together with any attributes added by the Jane Street compiler.

val fun_param : Ppxlib_ast.Asttypes.arg_label -> Ppxlib_ast.Parsetree.pattern -> function_param

fun_param lbl pat is Pparam_val (lbl, None, pat). This gives a more self-documenting way of constructing the usual case: value parameters without optional argument defaults.

val add_fun_param : ?attrs:Ppxlib_ast.Parsetree.attributes -> ?return_constraint:Ppxlib_ast.Parsetree.core_type -> Ppxlib_ast.Asttypes.arg_label -> Ppxlib_ast.Parsetree.expression option -> Ppxlib_ast.Parsetree.pattern -> Ppxlib_ast.Parsetree.expression -> Ppxlib_ast.Parsetree.expression

Say an expression is a "function" if it is a Pexp_fun or a Pexp_function. All functions have parameters and arity.

Suppose add_param lbl def pat e ==> e'. Then, letting param = Pparam_val (lbl, def, pat),

If e is a function with arity n, then e' is a function with arity n+1. param is added at the outermost layer. For example, if e = fun <params> -> <body>, then e' = fun <param :: params> -> body. The attributes on the resulting expression will be the attrs argument together with any attributes already present on e.

If e is not a function, then e' is a function with arity 1, namely: fun <param> -> <e>. The attributes of the resulting expression will be the attrs argument together with any attributes added by the Jane Street compiler.

val add_fun_params : ?attrs:Ppxlib_ast.Parsetree.attributes -> ?return_constraint:Ppxlib_ast.Parsetree.core_type -> function_param list -> Ppxlib_ast.Parsetree.expression -> Ppxlib_ast.Parsetree.expression

add_params params e is List.fold_right params ~init:e ~f:add_param. Note the fold_right: if e is fun <params'> -> <body>, then add_params params e is fun <params @ params'> -> <body>.

val coalesce_fun_arity : Ppxlib_ast.Parsetree.expression -> Ppxlib_ast.Parsetree.expression

This operation is a no-op, except as interpreted by the Jane Street compiler. If e is a function with arity n with an expression body that itself is a function with arity m, then coalesce_fun_arity e is a function of arity n + m.

You should usually call coalesce_fun_arity on metaquot fun expressions whose body may be a function, e.g.:

coalesce_fun_arity [%expr fun x y -> [%e possibly_function]]

Unboxed types

val ptyp_unboxed_tuple : (string option * Ppxlib_ast.Parsetree.core_type) list -> Ppxlib_ast.Parsetree.core_type

val pexp_unboxed_tuple : ?attrs:Ppxlib_ast.Parsetree.attributes -> (string option * Ppxlib_ast.Parsetree.expression) list -> Ppxlib_ast.Parsetree.expression

val pexp_record_unboxed_product : ?attrs:Ppxlib_ast.Parsetree.attributes -> (Astlib.Longident.t Ppxlib_ast.Asttypes.loc * Ppxlib_ast.Parsetree.expression) list -> Ppxlib_ast.Parsetree.expression option -> Ppxlib_ast.Parsetree.expression

val ppat_record_unboxed_product : ?attrs:Ppxlib_ast.Parsetree.attributes -> (Astlib.Longident.t Ppxlib_ast.Asttypes.loc * Ppxlib_ast.Parsetree.pattern) list -> Ppxlib_ast.Asttypes.closed_flag -> Ppxlib_ast.Parsetree.pattern

val pexp_unboxed_field : ?attrs:Ppxlib_ast.Parsetree.attributes -> Ppxlib_ast.Parsetree.expression -> Astlib.Longident.t Ppxlib_ast.Asttypes.loc -> Ppxlib_ast.Parsetree.expression

val ppat_unboxed_tuple : ?attrs:Ppxlib_ast.Parsetree.attributes -> (string option * Ppxlib_ast.Parsetree.pattern) list -> Ppxlib_ast.Asttypes.closed_flag -> Ppxlib_ast.Parsetree.pattern

val pexp_idx : ?attrs:Ppxlib_ast.Parsetree.attributes -> block_access -> unboxed_access list -> Ppxlib_ast.Parsetree.expression

Expression literals

val eint64_u : int64 -> Ppxlib_ast.Parsetree.expression

e.g. #42L

val eint32_u : int32 -> Ppxlib_ast.Parsetree.expression

e.g. #42l

val enativeint_u : nativeint -> Ppxlib_ast.Parsetree.expression

e.g. #42n

val efloat_u : string -> Ppxlib_ast.Parsetree.expression

e.g. #42.

Pattern literals

val pint64_u : int64 -> Ppxlib_ast.Parsetree.pattern

e.g. #42L

val pint32_u : int32 -> Ppxlib_ast.Parsetree.pattern

e.g. #42l

val pnativeint_u : nativeint -> Ppxlib_ast.Parsetree.pattern

e.g. #42n

val pfloat_u : string -> Ppxlib_ast.Parsetree.pattern

e.g. #42.

Layouts

val ptyp_alias : ?attrs:Ppxlib_ast.Parsetree.attributes -> Ppxlib_ast.Parsetree.core_type -> Ppxlib_ast.Asttypes.label Ppxlib_ast.Asttypes.loc option -> jkind_annotation option -> Ppxlib_ast.Parsetree.core_type

val ptyp_poly : ?attrs:Ppxlib_ast.Parsetree.attributes -> (string Ppxlib_ast.Asttypes.loc * jkind_annotation option) list -> Ppxlib_ast.Parsetree.core_type -> Ppxlib_ast.Parsetree.core_type

val ptyp_quote : ?attrs:Ppxlib_ast.Parsetree.attributes -> Ppxlib_ast.Parsetree.core_type -> Ppxlib_ast.Parsetree.core_type

val ptyp_splice : ?attrs:Ppxlib_ast.Parsetree.attributes -> Ppxlib_ast.Parsetree.core_type -> Ppxlib_ast.Parsetree.core_type

val pexp_newtype : ?attrs:Ppxlib_ast.Parsetree.attributes -> string Ppxlib_ast.Asttypes.loc -> jkind_annotation option -> Ppxlib_ast.Parsetree.expression -> Ppxlib_ast.Parsetree.expression

val type_declaration : ?attrs:Ppxlib_ast.Parsetree.attributes -> ?jkind_annotation:jkind_annotation -> name:string Astlib.Location.loc -> params: (Ppxlib_ast.Parsetree.core_type * (Ppxlib_ast.Asttypes.variance * Ppxlib_ast.Asttypes.injectivity)) list -> cstrs: (Ppxlib_ast.Parsetree.core_type * Ppxlib_ast.Parsetree.core_type * Astlib.Location.t) list -> kind:Ppxlib_ast.Parsetree.type_kind -> private_:Ppxlib_ast.Asttypes.private_flag -> manifest:Ppxlib_ast.Parsetree.core_type option -> unit -> Ppxlib_ast.Parsetree.type_declaration

Labeled tuples

val ptyp_tuple : ?attrs:Ppxlib_ast.Parsetree.attributes -> (string option * Ppxlib_ast.Parsetree.core_type) list -> Ppxlib_ast.Parsetree.core_type

val pexp_tuple : ?attrs:Ppxlib_ast.Parsetree.attributes -> (string option * Ppxlib_ast.Parsetree.expression) list -> Ppxlib_ast.Parsetree.expression

val ppat_tuple : ?attrs:Ppxlib_ast.Parsetree.attributes -> (string option * Ppxlib_ast.Parsetree.pattern) list -> Ppxlib_ast.Asttypes.closed_flag -> Ppxlib_ast.Parsetree.pattern

Immutable arrays

val ppat_array : ?attrs:Ppxlib_ast.Parsetree.attributes -> Ppxlib_ast.Asttypes.mutable_flag -> Ppxlib_ast.Parsetree.pattern list -> Ppxlib_ast.Parsetree.pattern

val pexp_array : ?attrs:Ppxlib_ast.Parsetree.attributes -> Ppxlib_ast.Asttypes.mutable_flag -> Ppxlib_ast.Parsetree.expression list -> Ppxlib_ast.Parsetree.expression