Module Ppxlib_jane.For_testing

module Language_extension : sig ... end

A ppxlib_jane-specific copy of the Jane Street-internal language extensions framework that reports that every language extension is enabled. This is the most permissive behavior possible, which is what we want for ppxes that process pieces of Jane Syntax.

module Language_extension_kernel : sig ... end

Language extensions provided by the Jane Street version of the OCaml compiler.