Module For_testing.Language_extension

A ppxlib_jane-specific copy of the Jane Street-internal language extensions framework that reports that every language extension is enabled. This is the most permissive behavior possible, which is what we want for ppxes that process pieces of Jane Syntax.

type nonrec 'a t

val is_enabled : _ t -> bool

Check if a language extension is currently enabled.

val is_at_least : 'a t -> 'a -> bool