Module type Language_extension_kernel.Language_extension_for_jane_syntax

This module type defines the pieces of functionality used by Jane_syntax_parsing and Jane_syntax so that we can more easily import these modules into ppxlib_jane, without also including all of the Language_extension machinery.

It includes the stateful operations that Jane_syntax_parsing relies on. This limits the number of bindings that ppxlib_jane needs to have mock implementations for.

type nonrec 'a t = 'a t

val is_enabled : _ t -> bool

Check if a language extension is currently enabled.

val is_at_least : 'a t -> 'a -> bool